Some Alignak specific views: livesynthesis, status, events log
