+++ etc/alignak.d/alignak-module-webui.ini
@@ -34,7 +34,7 @@ python_name=alignak_webui
 ;alignak_check_period = 10
 
 # The WebUI will store internally an events log queue limited to alignak_events_count items
-# If ALIGNAK_EVENTS_LOG_COUNT environment variable is set it will take précédence
+# If ALIGNAK_EVENTS_LOG_COUNT environment variable is set it will take precedence
 ;alignak_events_count = 1000
