+++ test/test-configurations/alignak/etc/alignak.d/alignak-module-ui-graphite2.ini
@@ -7,7 +7,7 @@ module_type=graphite-webui
 python_name=no_matter_what_is_set_here
 enabled=0
 
-uri=http://YOURSERVERNAME/
+uri=http://192.168.43.193/
                        ; Set your Graphite URI. Note : YOURSERVERNAME will be
                        ; changed by your broker hostname
