+++ test/test-configurations/alignak/etc/alignak.d/alignak-module-webui.ini
@@ -34,7 +34,7 @@ log_file=/tmp/var/log/alignak/shinken-webui.log
 ;alignak_check_period = 10
 
 # The WebUI will store internally an events log queue limited to alignak_events_count items
-# If ALIGNAK_EVENTS_LOG_COUNT environment variable is set it will take précédence
+# If ALIGNAK_EVENTS_LOG_COUNT environment variable is set it will take precedence
 ;alignak_events_count = 1000
 
 
@@ -113,7 +113,7 @@ logs_collection=alignak_events
 
 ## Declare the list of the WebUI modules
 ;modules=auth-ws-glpi
-modules=auth-ws-glpi, ui-graphite2, alignak-logs
+modules=ui-graphite2, alignak-logs
 # Specify the directory where to search for the WebUI modules
 # Default is /var/lib/shinken/modules because the Shinken installer locates the module in this directory
 # You would rather copy the Shinken modules in a dedicated place
