+++ test/test-configurations/alignak/etc/alignak.d/daemons.ini
@@ -20,6 +20,7 @@ name=arbiter-master
 ; ---
 ; My listening interface
 ;host=127.0.0.1
+host=0.0.0.0
 ; Set the host to 0.0.0.0 to allow Arbiter WS access from another system
 port=7770
 ; My adress for the other daemons
