+++ test/test-configurations/alignak/etc/alignak.d/modules.ini
@@ -101,7 +101,7 @@ enabled = 1
 
 ;
 ; Module specific parameters
-graphite_host=localhost
+graphite_host=192.168.43.193
 graphite_port=2004
 graphite_prefix=alignak
