+++ test/test-configurations/alignak/etc/alignak.ini
@@ -465,7 +465,9 @@ daemons_dispatch_timeout=2
 ; By default at localhost:8125 (UDP) with the alignak prefix
 ; --------------------------------------------------------------------
 ;statsd_host = localhost
+statsd_host=192.168.43.193
 ;statsd_port = 8125
+statsd_port=2004
 ;statsd_prefix = alignak
 ; --------------------------------------------------------------------
 ; When graphite_enabled is set, the Alignak internal metrics are sent
@@ -476,6 +478,7 @@ daemons_dispatch_timeout=2
 ; Default is not enabled for any interface
 ;statsd_enabled = 0
 ;graphite_enabled = 0
+graphite_enabled = 1
 ; --------------------------------------------------------------------
 
 
@@ -572,6 +575,7 @@ daemons_dispatch_timeout=2
 ; to its Graphite interface.
 ; See the alignak.d/modules.ini configuration file for more details / parameters
 ;host_perfdata_file=
+host_perfdata_file=test
 ;service_perfdata_file=
 ; --------------------------------------------------------------------
