+++ test/test-configurations/alignak/etc/arbiter/objects/hosts/mariadbs.cfg
@@ -23,3 +23,29 @@ define host{
    _MYSQLUSER           alignak
    _MYSQLPASSWORD       Alignak-2018
 }
+
+define host{
+   use                  mysql,freebsd-nrpe
+
+   host_name            mysql-master
+   alias                MySQL master
+   display_name         MySQL master
+
+   address              10.0.2.61
+
+   _MYSQLUSER           alignak
+   _MYSQLPASSWORD       Alignak-2018
+}
+
+define host{
+   use                  mysql-slave,freebsd-nrpe
+
+   host_name            mysql-slave
+   alias                MySQL slave
+   display_name         MySQL slave
+
+   address              10.0.2.62
+
+   _MYSQLUSER           alignak
+   _MYSQLPASSWORD       Alignak-2018
+}
