+++ test/test-configurations/alignak/etc/arbiter/objects/hosts/nginxs.cfg
@@ -1,0 +1,33 @@
+define host{
+   use                  nginx,freebsd-nrpe
+
+   host_name            my-nginx-rp1
+   alias                Nginx server
+   display_name         Nginx rp1
+
+   address              10.0.2.2
+
+   ; Default status URL
+   _PORT               80
+   _STATUS_URL         status
+   _SERVER_NAME        localhost
+}
+
+define host{
+   use                  nginx-auth,freebsd-nrpe
+
+   host_name            my-secured-nginx-es2
+   alias                Nginx server - HTTPS
+   display_name         Nginx rp2
+
+   address              10.0.2.3
+
+   ; Default status URL
+   _PORT               80
+   _STATUS_URL         status
+   _SERVER_NAME        localhost
+
+   ; Default username / password
+   _USERNAME           admin
+   _PASSWORD           admin
+}
