+++ test/test-configurations/alignak/etc/arbiter/packs/nginx/commands.cfg
@@ -1,0 +1,12 @@
+# -----------------------------------------------------------------
+# Default ES commands
+# -----------------------------------------------------------------
+define command {
+    command_name    check_nginx_status
+    command_line    $LIBEXEC$/check_nginx_status.pl -H $HOSTADDRESS$ -P $HOSTPORT$ -u $_HOSTSTATUS_URL$ -s $_HOSTSERVER_NAME$ -w $_HOSTACTIVE_CONN_WARN$,$_HOSTREQ_PER_SEC_WARN$,$_HOSTCONN_PER_SEC_WARN$ -c $_HOSTACTIVE_CONN_CRIT$,$_HOSTREQ_PER_SEC_CRIT$,$_HOSTCONN_PER_SEC_CRIT$
+}
+# Check with authentication
+define command {
+    command_name    check_nginx_status_auth
+    command_line    $LIBEXEC$/check_nginx_status.pl -H $HOSTADDRESS$ -P $HOSTPORT$ -u $_HOSTSTATUS_URL$ $_HOSTSERVER_NAME$ -S -U $_HOSTUSERNAME$ -P $_HOSTPASSWORD$ -w $_HOSTACTIVE_CONN_WARN$,$_HOSTREQ_PER_SEC_WARN$,$_HOSTCONN_PER_SEC_WARN$ -c $_HOSTACTIVE_CONN_CRIT$,$_HOSTREQ_PER_SEC_CRIT$,$_HOSTCONN_PER_SEC_CRIT$ --ssl
+}
