+++ test/test-configurations/alignak/etc/arbiter/packs/nginx/groups.cfg
@@ -1,0 +1,10 @@
+# Host and service groups
+define hostgroup {
+    hostgroup_name      nginx
+    alias               Nginx hosts
+}
+
+define servicegroup {
+    servicegroup_name   nginx
+    alias               Nginx services
+}
