+++ test/test-configurations/alignak/etc/arbiter/packs/nginx/templates.cfg
@@ -1,0 +1,57 @@
+# Nginx host
+# ----------
+define host {
+    name                nginx
+    register            0
+
+    hostgroups          nginx
+
+    ; Default status URL
+    _PORT               80
+    _STATUS_URL         nginx_status
+    _SERVER_NAME        localhost
+
+    _ACTIVE_CONN_WARN   500
+    _ACTIVE_CONN_CRIT   1000
+    _REQ_PER_SEC_WARN   50
+    _REQ_PER_SEC_CRIT   75
+    _CONN_PER_SEC_WARN  10
+    _CONN_PER_SEC_CRIT  20
+}
+
+define host {
+    name                nginx-auth
+    register            0
+
+    hostgroups          nginx
+
+    ; Default status URL
+    _STATUS_URL         nginx_status
+    _SERVER_NAME        localhost
+
+    ; Default username / password
+    _USERNAME           admin
+    _PASSWORD           admin
+}
+
+define service {
+    service_description check_status
+    register            0
+    host_name	         nginx
+
+    check_command       check_nginx_status
+
+    aggregation         nginx
+    servicegroups       nginx
+}
+define service {
+    service_description check_status_auth
+    alias               Check status (authenticated)
+    register            0
+    host_name	         nginx-auth
+
+    check_command       check_nginx_status_auth
+
+    aggregation         nginx
+    servicegroups       nginx
+}
