+++ test/test-configurations/alignak/etc/arbiter/templates/generic-host.cfg
@@ -15,6 +15,8 @@ define host{
     active_checks_enabled     1
     check_period              24x7
 
+    _GRAPHITE_PRE               alignak
+
     # Notification part
     # One notification each day (1440 = 60min* 24h)
     # every time, and for all 'errors'
